requirements.txt
asttokens==2.0.5
backcall==0.2.0
backports.functools-lru-cache==1.6.4
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
debugpy==1.6.0
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
executing==0.8.3
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.15.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.23.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
pip==21.2.2
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
Pygments==2.12.0
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==23.1.0
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
setuptools==61.2.0
six==1.16.0
sklearn==0.0
stack-data==0.3.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
wheel==0.37.1
wincertstore==0.2
zipp==3.8.0
